+++ packages/at_commons/lib/src/verb/operation_enum.dart
@@ -27,9 +27,9 @@ enum EnrollOperationEnum {
   deny,
   revoke,
   list,
-  update,
   fetch,
-  unrevoke
+  unrevoke,
+  delete
 }
 
 String getEnrollOperation(EnrollOperationEnum? enrollOperationEnum) =>