+++ packages/at_utils/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.0.19
+- build[deps]: Upgraded the following packages:
+  - at_commons to v5.0.0
+## 3.0.18
+- chore: Upgrade at_commons to 4.1.1
+## 3.0.17
+- chore: Upgrade at_commons to 4.1.0
 ## 3.0.16
 - fix: Uptake at_commons 4.0.0
 ## 3.0.15